+++ NEWS.md
@@ -1,6 +1,7 @@
 
 # ranger 0.15.4
 * Add node.stats option to save node statistics of all nodes
+* Add horizontal shrinkage
 
 # ranger 0.15.3
 * Fix min bucket option in C++ version